CG-133 Refactor: ...button 들 전부 app/custom-components/buttons/ 로 이동시킴
- 그 외, 누락되거나 다른곳에 있던 이미지들 정리해서 images 에 추가함
