+++ app/custom-components/basics/row/row.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { View, ViewProps } from "react-native"
-import { styles } from "./styles"
+import { styles } from "../common-styles"
 //- TODO: RowProps 생성 (ScreenProps) 참고할 것
 // export const Row = (props: RowProps) => {
 export const Row = (props: ViewProps) => {
