+++ app/custom-components/basics/view-component/view-component.tsx
