+++ app/custom-components/buttons/caregiver-type-button/caregiver-type-button.tsx
@@ -1,7 +1,7 @@
 import { View, Text } from "react-native"
 import React from "react"
 import { styles } from "./styles"
-import { PreBol12 } from "../custom-texts/custom-texts"
+import { PreBol12 } from "../../custom-texts/custom-texts"
 import { palette } from "~/app/theme"
 
 const CaregiverTypeButton = ({ text, style }: { text: string; style?: Object }) => {
