+++ app/custom-components/caregiver-certificate/caregiver-certificate.tsx
@@ -4,7 +4,7 @@ import { styles } from "./styles"
 import { HEIGHT, WIDTH } from "~/app/theme"
 import { PreReg12, PreReg16 } from "../custom-texts/custom-texts"
 import { SUB_HEAD_LINE } from "~/app/theme/palette"
-import { Row } from "../boxes/basics/row"
+import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
 
 export const CaregiverCertificate = (props) => {
