+++ app/custom-components/caregiver-name-star-review/caregiver-name-star-review.tsx
@@ -4,7 +4,7 @@ import { styles } from "./styles"
 import { WIDTH, HEIGHT } from "~/app/theme"
 import { PreBol14, PreMed16 } from "../custom-texts/custom-texts"
 import { DBG, GIVER_CASUAL_NAVY, HEAD_LINE, SUB_HEAD_LINE } from "~/app/theme/palette"
-import { Row } from "../boxes/basics/row"
+import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
 import { DivisionLineVertical } from "../division-line-vertical/division-line-vertical"
 import { navigate } from "~/app/navigators"
