+++ app/custom-components/comment/comment.tsx
@@ -4,7 +4,7 @@ import { styles } from "./styles"
 import { HEIGHT, WIDTH } from "~/app/theme"
 import { PreReg12, PreReg14 } from "../custom-texts/custom-texts"
 import { BODY, HEAD_LINE, LIGHT_LINE, SUB_HEAD_LINE } from "~/app/theme/palette"
-import { Row } from "../boxes/basics/row"
+import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
 import { DivisionLine } from "../division-line/division-line"
