+++ app/custom-components/comment/styles.ts
@@ -1,7 +1,5 @@
-import { GIVER_CASUAL_NAVY, CARE_NATURAL_BLUE } from "~/app/theme/palette"
 import { StyleSheet } from "react-native"
 import { WIDTH, HEIGHT } from "~/app/theme"
-import { BASIC_BACKGROUND_PADDING_WIDTH } from "../view-component/view-component"
 
 export const styles = StyleSheet.create({
   root: {
