+++ app/custom-components/dropdown-boxes/pet-info-dropdown-box/pet-info-dropdown-box.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import { FlatList } from "react-native-gesture-handler"
 import { PetProfileCard } from "../../pet-profile-card/pet-profile-card"
 import { styles } from "./styles"
-import { Row } from "../../boxes/basics/row"
+import { Row } from "../../basics/row/row"
 import { PopSem14 } from "../../custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { PetInfoDropdownBoxProps } from "./pet-info-dropdown-box.props"
