+++ app/custom-components/dropdown-boxes/select-option-button/select-option-dropdown-box.tsx
@@ -1,6 +1,6 @@
 import { View, Image, Pressable } from "react-native"
 import React from "react"
-import { Row } from "../../boxes/basics/row"
+import { Row } from "../../basics/row/row"
 import { PreReg12 } from "../../custom-texts/custom-texts"
 import { FlatList } from "react-native-gesture-handler"
