+++ app/custom-components/index.ts
@@ -2,12 +2,14 @@
 export * from "./custom-texts/custom-texts"
 
 //* buttons
+export * from "./buttons/caregiver-type-button/caregiver-type-button"
 export * from "./buttons/pressable-button/pressable-button"
 export * from "./buttons/come-home-go-to-switch-button/come-home-go-to-switch-button"
 export * from "./buttons/conditional-button/conditional-button"
 export * from "./buttons/row-rounded-button/row-rounded-button"
 export * from "./buttons/public-private-switch-button/public-private-switch-button"
 export * from "./buttons/make-booking-button/make-booking-button"
+export * from "./buttons/sitter-profile-button/sitter-profile-button"
 
 //* headers
 export * from "./service-type-indicator-header/service-type-indicator-header"
@@ -16,11 +18,11 @@ export * from "./screen-headers/home-screen-header/home-screen-header"
 export * from "./screen-headers/all-comments-screen-header/all-comments-screen-header"
 
 //* others
-export * from "./view-component/view-component"
-export * from "./service-choice/service-choice-button/service-choice-button"
-export * from "./sitter-profile/sitter-profile-button/sitter-profile-button"
-export * from "./boxes/basics/row"
-export * from "./boxes/basics/row-rounded-box"
+export * from "./basics/view-component/view-component"
+export * from "./buttons/service-choice-button/service-choice-button"
+export * from "./buttons/sitter-profile-button/sitter-profile-button"
+export * from "./basics/row/row"
+export * from "./basics/row-rounded-box/row-rounded-box"
 export * from "./division-line/division-line"
 export * from "./dots-indicator/dots-indicator"
 export * from "./selected-pet-card/selected-pet-card"
