+++ app/custom-components/row-rounded-time-interval-picker/row-rounded-time-interval-picker.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react"
 import { Image, Pressable, Text, View, ViewProps } from "react-native"
 import { styles } from "./styles"
-import { RowRoundedBox } from "../boxes/basics/row-rounded-box"
 import { PopReg16, PopSem16, PreReg16 } from "../custom-texts/custom-texts"
 import RNDateTimePicker from "@react-native-community/datetimepicker"
-import { Row } from "../boxes/basics/row"
+import { Row } from "../basics/row/row"
 import { WIDTH, HEIGHT } from "~/app/theme"
+import { RowRoundedBox } from "../basics/row-rounded-box/row-rounded-box"
 
 export const RowRoundedTimeIntervalPicker = (props) => {
   const { style, platform } = props
