+++ app/custom-components/select-pet-item/select-pet-item.tsx
@@ -3,7 +3,7 @@ import React, { useState, useLayoutEffect } from "react"
 import { styles } from "./styles"
 import { PreBol16, PreReg14 } from "../custom-texts/custom-texts"
 import { BODY } from "~/app/theme/palette"
-import { Row } from "../boxes/basics/row"
+import { Row } from "../basics/row/row"
 import { BlueCheckbox } from "../blue-checkbox/blue-checkbox"
 
 export const SelectPetItem = (props) => {
